ai-service/requirements.txt
flask
flask-cors
sentence-transformers
scikit-learn
numpy
pymupdf
python-docx
pydantic
